Player Attack + Enemy Turn
Applied sprites to player and enemies. Three enemies will now spawn at random locations on the map (to be used in the future). The player can still use WASD to move during their turn, but as soon as they do, the other enemies will move in a random available direction (if possible) at the same time. The player can also use the arrow keys to attack in a direction, taking up their turn (messages with damage/health values are displayed in the log for now).